Add GUI for PSU packing
